# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Startup de Bolso API",
        "description": "Documentação da API do painel de agendamentos e gestão de usuários da Startup de Bolso.",
        "version": "1.0.0",
        "contact": {
            "email": "suporte@startupdebolso.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o login e recebe um JWT",
                "security": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token gerado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Retorna o usuário autenticado",
                "responses": {
                    "200": {
                        "description": "Dados do usuário"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Invalida o token JWT",
                "responses": {
                    "200": {
                        "description": "Deslogado com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/password/email": {
            "post": {
                "tags": [
                    "Recuperação de Senha"
                ],
                "summary": "Envia o código de 6 dígitos via e-mail",
                "security": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "cliente@email.com"
                                    }
                                },
                                "required": [
                                    "email"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "E-mail enviado se o usuário existir"
                    }
                }
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "Recuperação de Senha"
                ],
                "summary": "Redefine a senha utilizando o código",
                "security": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "cliente@email.com"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nova_senha123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "nova_senha123"
                                    }
                                },
                                "required": [
                                    "email",
                                    "code",
                                    "password",
                                    "password_confirmation"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada com sucesso"
                    },
                    "400": {
                        "description": "Código inválido ou dados incorretos"
                    }
                }
            }
        },
        "/events": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Lista todos os agendamentos do usuário",
                "responses": {
                    "200": {
                        "description": "Lista de eventos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Cria um novo agendamento",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Reunião de Alinhamento"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "2026-03-05 14:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "2026-03-05 15:00:00"
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "João Alencar"
                                    },
                                    "event_type": {
                                        "type": "string",
                                        "example": "venue_visit"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "blue"
                                    }
                                },
                                "required": [
                                    "title",
                                    "start_time",
                                    "end_time"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento criado"
                    },
                    "422": {
                        "description": "Conflito de horário ou erro de validação"
                    }
                }
            }
        },
        "/events/{id}": {
            "put": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Atualiza um agendamento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "string"
                                    },
                                    "end_time": {
                                        "type": "string"
                                    },
                                    "customer_name": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Exclui um agendamento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento deletado"
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin - Usuários"
                ],
                "summary": "Lista todos os usuários",
                "responses": {
                    "200": {
                        "description": "Operação feita apenas por admins"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin - Usuários"
                ],
                "summary": "Cria um novo usuário ou admin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado"
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "put": {
                "tags": [
                    "Admin - Usuários"
                ],
                "summary": "Atualiza os dados de um usuário",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salvo"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin - Usuários"
                ],
                "summary": "Excluiqum",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excluido"
                    }
                }
            }
        }
    }
}